// src/client/evm-rpc.ts
export interface EvmLog {
  /** Address from which this log originated */
  address: string
  /** Array of 0 to 4 32 Bytes DATA of indexed log arguments */
  topics: string[]
  /** Contains non-indexed arguments of the log */
  data: string
  /** Block number where this log was in (hex) */
  blockNumber: string
  /** Hash of the transaction this log was created from */
  transactionHash: string
  /** Index position of the transaction in the block (hex) */
  transactionIndex: string
  /** Hash of the block where this log was in */
  blockHash: string
  /** Index position in the block (hex) */
  logIndex: string
  /** true when the log was removed due to a chain reorg */
  removed: boolean
}